1. created enums file 2. abstracted commonly used functions to utility file
