+++ python/enums.py
@@ -1,0 +1,6 @@
+YEARS = ['2017', '2018', '2019', '2020', '2021']
+INTERVALS = ["1m", "5m", "15m", "30m", "1h", "2h", "4h", "6h", "8h", "12h", "1d", "3d", "1w", "1mo"]
+DAILY_INTERVALS = ["1m", "5m", "15m", "30m", "1h", "2h", "4h", "6h", "8h", "12h", "1d"]
+MONTHS = list(range(1,13))
+MAX_DAYS = 35
+BASE_URL = 'https://data.binance.vision/'
